msm: camera: sync: Add warning logs
This change checks the state of the sync object, it's parents
and children before detroying the sync obj. If it finds any of
above mentioned objects to be in active state it will print a
warning message.

Change-Id: I4966c6c80f751e8cb41f7f10e4493f9599677d62
Signed-off-by: Karthik Anantha Ram <kartanan@codeaurora.org>
(cherry picked from commit 9df75e312ca01894fc12ed29090f1fea9a3b9459)