Revert "fib_rules: return 0 directly if an exactly same rule exists when NLM_F_EXCL not supplied"
This reverts commit 1fff19a925e524556e85efcd728efad5274ce5b6.

Some users are having soft reboots with loss of connectivity because
netd does not like this change. I would report this upstream for
verification but Android is always special and I don't currently have
the time to help debug or get a fix tested (especially if I can't
reproduce on my current setup).

Change-Id: I90b8d3b15c3f4bed123995bca648173427e6eb3f
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>